+++ projects/demo/src/app/demo/demo.module.ts
@@ -118,6 +118,20 @@ import { DistributePageComponent } from './pages/distribute-page/distribute-page
 import { CardPageComponent } from './pages/card-page/card-page.component';
 import { RouterModule } from '@angular/router';
 import { DatePickerPageComponent } from "./pages/date-picker-page/date-picker-page.component";
+import {
+    CommonTableDescriptionComponent
+} from './pages/table-page/common-table-description/common-table-description.component';
+import {
+    TableSortableHeaderPageComponent
+} from './pages/table-page/table-sortable-header-page/table-sortable-header-page.component';
+import { TableRowPageComponent } from './pages/table-page/table-row/table-row-page.component';
+import { TableCellPageComponent } from './pages/table-page/table-cell-page/table-cell-page.component';
+import {
+    TableSortableHeaderCellPageComponent
+} from './pages/table-page/table-sortable-header-cell-page/table-sortable-header-cell-page.component';
+import {
+    TableLeadCellMultiLinePageComponent
+} from './pages/table-page/table-lead-cell-multi-line-page/table-lead-cell-multi-line-page.component';
 
 const COMPONENTS = [
     DemoComponent,
@@ -202,6 +216,12 @@ const COMPONENTS = [
     ScrollbarPageComponent,
     PopoverPageComponent,
     DistributePageComponent,
+    CommonTableDescriptionComponent,
+    TableRowPageComponent,
+    TableCellPageComponent,
+    TableSortableHeaderPageComponent,
+    TableSortableHeaderCellPageComponent,
+    TableLeadCellMultiLinePageComponent,
 ];
 
 @NgModule({