+++ projects/demo/src/app/demo/pages/table-page/common-table-description/common-table-description.component.ts
@@ -1,0 +1,10 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+    selector: 'pa-demo-common-table-description',
+    templateUrl: './common-table-description.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class CommonTableDescriptionComponent {
+
+}