+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # 2.50.0 (2022-05-20)
 
 ### Breaking changes:
-- **Table sortable header**: viewport mode as input instead of using directly breakpoint observer [mpellerin42]
+- **Table sortable header**: viewport mode is now an input instead of coming directly from breakpoint observer [mpellerin42]
+
+### Improvements:
+- Better documentation for table components and directives [mpellerin42]
 
 # 2.49.1 (2022-05-18)